Testing the updated version of Floreada, first glance the named list() is gone but no transformation is applied to the plots.
Glance at the unopened .wsp, Comp- application not present, edited version restored parse ability similar last round.

Additional noticed my edited version differs from opened transforms in not showing Comp- in front of the gettransformations.

Need to look into this a bit more, with an uncomped file to make sure we are not messing with something we will need to reverse in a different setting.